requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Babel==4.0.0
qrcode==7.4.2
Pillow==10.0.1
Werkzeug==3.0.0
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.22
python-dotenv==1.0.0
typing-extensions==4.8.0
blinker==1.6.3
greenlet==3.0.0
pytz==2023.3.post1
Babel==2.13.0
pip-review
